1. next.config.js: pg-cloudflare disabled on client side. 2. utils/db/repos contains the function to query list of repo-db objects from a list of repo names3. Updated the pg library because of bugs in the previous version
