+++ utils/db/repos.ts
@@ -1,6 +1,18 @@
 import conn from '.';
-import type { RepoIdentifier } from '../../types/repository';
+import type { DbRepo, RepoIdentifier } from '../../types/repository';
 import { convert } from './converter';
+
+export const getReposFromNames = async (allRepos: RepoIdentifier[]) => {
+	const allReposFormattedAsTuples = allRepos.map(repo => `(${convert(repo.repo_provider)}, ${convert(repo.repo_owner)}, ${convert(repo.repo_name)})`).join(',');
+	const repo_list_q = `SELECT *
+		FROM repos 
+		WHERE (repo_provider, repo_owner, repo_name) IN (${allReposFormattedAsTuples})`;
+	const result: { rows: DbRepo[] } = await conn.query(repo_list_q).catch(err => {
+		console.error(`[RepoList] Error in getting repository-list from the database`, err);
+		throw Error(err); // FIXME: handle this more elegantly
+	});
+	return result.rows
+}
 
 export const setRepoConfig = async (repo: RepoIdentifier, configType: 'auto_assign' | 'comment', value: boolean) => {
 	const update_repo_config_q = `UPDATE repos
